Handle brutal disconnect

+++ DeusCore/TcpSocket.h
@@ -23,22 +23,22 @@ namespace DeusCore
 		void TCPConnect(const SOCKET socket, const bool setNonBlocking = false);
 		
 		// Send buffer to the connected socket
-		bool TCPSend(const Packet& packet, size_t& byteSent);
+		bool TCPSend(const Packet& packet, int& byteSent);
 
 		// Send buffer to the connected socket
-		bool TCPSend(const Buffer512& buffer, size_t& byteSent);
+		bool TCPSend(const Buffer512& buffer, int& byteSent);
 
 		// Send buffer to the connected socket
-		bool TCPSend(const char *data, size_t size, size_t &byteSent);
+		bool TCPSend(const char *data, size_t size, int &byteSent);
 
 		// Receive ONLY the first packet (and delete other datas) from the connected socket
-		bool TCPRecv(std::unique_ptr<Packet>& p_packetReceived, size_t& byteRecv) ;
+		bool TCPRecv(std::unique_ptr<Packet>& p_packetReceived, int& byteRecv) ;
 
 		// Receive buffer from the connected socket
-		bool TCPRecv(Buffer512& buffer, size_t& byteRecv) ;
+		bool TCPRecv(Buffer512& buffer, int& byteRecv) ;
 
 		// Receive informations from the connected socket
-		bool TCPRecv(char *data, size_t size, size_t &byteRecv) ;
+		bool TCPRecv(char *data, size_t size, int &byteRecv) ;
 	
 		
 	};
